Updated openstack/openstack Project: openstack-infra/project-config  a7e79abf6ad783967a1d45213784aa03e65250bb
Make neutron functional job non-voting

It currently blocks merging patches due to multiple instability issues
that are most likely not related to neutron code, and leaving it this
way so close to release is unsafe.

This patch is expected to revert asap when we regain control of the job
state.

Change-Id: Ib14ee3cdd9ea0004f5e50204b72819c1c5c54367
Related-Bug: #1707933
Related-Bug: #1709032