Updated openstack/openstack Project: openstack/nova  4dc7682e7979c9bdb96dce78863ce9b29e863e34
placement: filtering the resource provider id when delete trait association

When delete a trait association for specific resource provider, the current
code will delete that trait associations for all the resource providers.

The expected behaviour is only the specific trait association for the
specific resource provider will be removed. This patch corrects the filter
for the deletion of trait association.

Change-Id: I4b8f1b8f0a8d16f1c415edbfb2a7d342d176f70a
Closes-Bug: #1708978