Updated openstack/openstack Project: openstack/tempest  8e3296552cba01fbe85eb9387deeccadb7a3b2af
Fix verify config API version checks

The verify config tool checks that tempest configuration matches
the discovered versions, but it uses clients for a specific version
to make the discovery, which may not work if that version is
configured as not available.

This patch makes the code robust to those cases so that it may work
with different configurations and target clouds.

Change-Id: I65e2a685f683c0a3f141cb5349abc6065143bcdc