Updated openstack/openstack Project: openstack/nova  c780f31c795ee9a86e61419f1fbd79a0d5a0b945
Handle ironicclient failures in Ironic driver

The ironic driver uses ironicclient to get the list of nodes from the
Ironic service. If that service is not running, or something otherwise
prevents the ironicclient's call from succeeding, and unhandled
exception will occur. This patch adds a generic exception handler for
this call, logging an error and continuing.

Co-Authored-By: Lucas Alvares Gomes <lucasagomes@gmail.com>

Closes-Bug: #1706772

Change-Id: I4ebebcf221dfe29e2aa125f5956aec10108f8fbe