Updated openstack/openstack Project: openstack/tempest  4207dbedad4db343f86e9e8fe4e81db1a5012694
Make verify_tempest_config workspace aware

Hard coded relative path to the source code is changed, so that
the tool is workspace aware.
Edited help for -o argument. When -u is defined and -o not, description of -o
argument was misleading because the new config file was not printed to STDOUT.
The patch adds documentation of tool's options. Environment variables are
documented as well.

Change-Id: I31fef80fb98f22711638908bef3a17bf18d63772
Closes-Bug: #1683809