Updated openstack/openstack Project: openstack/nova  fa3f2e47965ce4d51d2f42b1d4a09ca735539766
Add functional test for local delete allocations

This adds a test that shows we clean up allocations for an instance
that was local deleted while the compute host was down, when the
compute host comes back up.

There might still be a problem if the compute host is never brought
back up, as allocations will still exist for the instance and show
up as usage during usage queries to placement.

Related-Bug: #1679750

Change-Id: Ia68a5a69783963b063264edde84006973bb77ceb