Updated openstack/openstack Project: openstack/barbican  e6c4f8cd7363f6662c30b276140fb2b8c9132326
Revert "Use devstack functions for deploying barbican-svc"

Castellan unintentionally can't handle a barbican URL that has a path in
addition to the hostname, such as http://ip-address/key-manager, unless
it is followed by a forward slash (http://ip-address/key-manager/ ).  We
should either revert this change before rc1 or merge
https://review.openstack.org/#/c/491942/, make a new release of
Castellan, and beg for a change in upper-constraints for castellan to
handle the new release.

This reverts commit 508a34e23c05013a7ba1f33120c78e0da5cc8f28.

Change-Id: Iceb3a5fa890d64468cd6e7f5dec297d11a274d20