Updated openstack/openstack Project: openstack/cinder  4b45a359fb0d05ac752b8275fde9c388bd7b6d7a
cinder backup sets incorrect parent_id

If --incremental flag is used and an incremental backup
fails, it should update the parent_id of the backup to
None.

Also for ceph driver, any backup after the first one
is an incremental backup automatically. Hence, the
is_incremental field for cinder backup-show <backup-id>
is misleading since it is set to False currently.

The patch correctly updates the parent_id for a
backup in case the incremental backup is successful.
If not, then it retains the parent_id as None. This
will also update the parent_id to None if incremental
backup fails even when --incremental flag is used.

Change-Id: I22deff68d9376c2ecc88bd2d9d0e1b445d793123
Closes-Bug: #1703013