Updated openstack/openstack Project: openstack/keystone  2fa4169b60e57e00a0d7b9ca3ac5c3ffe8c0dd6c
Make an error state message more explicit

Make the error state message more explicit when the admin_token_auth
middleware is still in the paste-ini. This is an error, the operator
must remove the elements from their keystone paste ini or the deletion
of the code (in Queens release) will result keystone to fail to run.

This is an error in the config. If the operator is using the default
paste-ini form keystone, this error should not occur in Pike. If
the paste-ini is managed via CMS (e.g. Ansible) it is likely the
config must be changed.

Change-Id: Id40a7161574c879aa5c948b94173308a03d4c240