Updated openstack/openstack Project: openstack/os-net-config  e1c9456768a75c07f434ea527bab447062de99b7
This patch adds initial support for the Contrail vRouter interface

The patch introduces the Contrail vRouter interface object. It can be
configured as a kernel module or in dpdk mode. The parameters are:
bind_int (required): the interface the vRouter interface is bound to
dpdk (optional): boolean if dpdk or not

Closes bug: 1709862

Change-Id: Iecb739e3964bf5307b0605ea1d80d2369f60918b