Updated openstack/openstack Project: openstack/nova  e62fe1accf5f33cc6971d5b5c194beb846b9a528
doc: provide more details on scheduling with placement

This provides a more detailed, but still high level, explanation
of how the FilterScheduler is using allocation candidates to pick
hosts during scheduling, and how the allocations are handled during
a move operation, including a resize to the same host.

Ideally this content will live somewhere else in the devref, or
would be updated in the spec for blueprint placement-claims, but
for now this should suffice for Pike release notes that can point
at this upgrade doc.

Change-Id: I274c684f829bc310ebb17faabf498d36f4ceea0c