Updated openstack/openstack Project: openstack/nova  1ce3d78982a9c7f390406a5e79b8fcda5e61470d
Cleanup release note about ignoring allow_same_net_traffic

A large chunk of the release note for this is duplicated with
the same reno from e5080c733076f5098f85952051878f41d47f8181
but the formatting and wording is slightly different.

This fixes them to look similar for the duplicate text.

Change-Id: Idd6149ae85ac2724633b80e938c4c1bf981b772b