Updated openstack/openstack Project: openstack-infra/system-config  057c867895a3fe608a2bbf920981ea28cb0ad90f
Add images.linuxcontainers.org reverse proxy

In order to aid in the successful testing of
OpenStack-Ansible deployments which make use
of the upstream base containers, this adds a
reverse proxy for the LXC container images
and the API used by the LXC tooling to find
the right image.

The URL http://images.linuxcontainers.org
redirects to either the US or UK site based
on geolocation. To prevent getting caught in
the redirect we pick the US source.

Note that the API used by the LXc tooling
queries [1] in order to find the latest images,
then fetches that image via the path specified
in the last ; delimited field.

[1] http://images.linuxcontainers.org/meta/1.0/index-system
Change-Id: I8336aeb85689cfaac8a794a5c3f5c62273d8ded0