Updated openstack/openstack Project: openstack-infra/project-config  b193c08bbfeead2e94d1947ce5dacf62b1c64f47
Switch vexxhost to RAW image format

The VEXXHOST public cloud uses Ceph-backed instances
for it's infrastructure.  When using QCOW2 images,
the content of the image is imported on every boot
of the instance, in some cases almost ~15GB write.

When a RAW image is uploaded, OpenStack Nova will
use copy on write to create a new copy of the image
very quickly and efficently, speeding up boot times
and reducing load.

Change-Id: I0210b70b6812588f3e5dcad10c08d4ee8be29be3