Updated openstack/openstack Project: openstack/nova  35cf533e12c07dc2c68fd2efbbed3cb2c3cb896f
Add release note for shared storage known issue

Due to the scheduler report client on the compute service
not being aware of shared storage providers in Placement
for a given compute node, we can't claim support for
shared storage pool modeling in Pike. This adds a release
note to be clear about the lack of that functionality.

Change-Id: I0692a4fdf9a19889c5d708fb9a6a562c914963f9
Related-Bug: #1707256