Updated openstack/openstack Project: openstack-infra/project-config  0818faa2229507fa363971a5b94e4fd516af6250
Set max-servers down to 0 for VEXXHOST

There is a bug with Glean where it can not handle config drives
which use the `vfat` format, as they come with a label of `CONFIG-2`
and not `config-2` which is causing it not to properly run on
the VEXXHOST cloud.

This patch disables the cloud until a patch and a new release of
Glean can be pushed out to address that issue.

Change-Id: Icc2a96f8a6303d458c3697cc14e0286b3fa55068