Updated openstack/openstack Project: openstack/nova  6c13b2d53814e0b64f1e9998484e3aed0a4b0320
Test resize with too big flavor

This test tries to resize to a flavor that requests more VCPU than
what the compute hosts has available and expect the resize to fail.
However placement returns no allocation candidate and the scheduler
falls back to the legacy filtering. As CoreFilter is not enabled the
filtering result in hosts selected without enough VCPU resource.

This is a test for an open bug therefore it asserts the wrong behavior
but it has the proper assert commented out.

Related-bug: #1708637
Change-Id: Ibd7f695d51b1f8497ca492968c7e5c69ada0a85c