Updated openstack/openstack Project: openstack/nova  c75a92e17916864a1ec48025a94b7bfa5b29f145
Raise NoValidHost if no allocation candidates

Placement took over the role of the CoreFilter, RamFilter and DiskFilter
from the FilterScheduler. Therefore if placement returns no allocation
candidates for a request then scheduling should be stopped as this means
there is not enough VCPU, MEMORY_MB, or DISK_GB available in any compute
node for the request.

Change-Id: If20a20e5cce7ab490998643e32556a1016646b07
Closes-Bug: #1708637