Updated openstack/openstack Project: openstack/ironic-python-agent  a48712018724eb9c24d41f901ed411a9f3cd4806
BMC can be configured using different lan channel

It may happen that BMC is configured to use non-zero channel.
In this case we should iterate across all of them as long as we
get a correct IP address (in this case different than "0.0.0.0"
which is a placeholder for not configured console).

Change-Id: I3c351af1882b24c8f56e4363249b19b5c3a4a446
Closes-Bug: #1702514