Updated openstack/openstack Project: openstack/nova  8ce1718b4dc9ef8b1e49b4c97573a2aa50ccf149
api-ref: requested security groups are not applied to pre-existing ports

When creating a server, you can request security groups and
pre-existing ports, but the security groups are only applied to
any new ports that nova creates, not the pre-existing ones that
the user passes in. This change makes a note of that in the API
reference.

Change-Id: I0ea6891f10f021b0ed752200417e87d7e9bfda31
Related-Bug: #1707319