Updated openstack/openstack Project: openstack/ironic  3d8626dc10fb69809f81839abae439ebb612194c
Add a flag to always perform persistent boot on PXE interface

With the proposal to solve the bug that PXE interface should
be using non-persistent boot for cleaning/deploying phases as
default(commit: c7091fb8e2898e526f8ce242f50096a2cabeb1fa), it's
necessary to create a flag to enable persistent behavior when
you set the boot device.
This flag will override a non-persistent behavior in the
cleaning and deploy process.

Change-Id: I1f47393c17a3f5319fffd963ec0a016b41865c5d
Closes-Bug: 1703945
Co-Authored-By: Stenio Araujo <steniaraujo@lsd.ufcg.edu.br>