Updated openstack/openstack Project: openstack/nova  9956333b560fd374a0d338b426b036ee1c5ff4ca
placement: refactor healing of allocations in RT

Updates the RT method for "healing" instance allocations so that we
only delete allocations (and delete them entirely for an instance) in
the event that the instance has been local-deleted. Adds log statements
to the method to give us a good idea of what the state of various
things are during test runs.

Note that this patch only partially fixes LP #1707071 to remove some of
the back-and-forth allocation trampling that was going on for move
operations. The next patch adds the final piece of the puzzle which is
to have the confirm/revert resize step on the source or destination
compute host *correct* a doubled-up allocation. What this patch does is
prevent the healing function from deleting allocations unless the
instance is local-deleted.

Partial-bug: #1707071

Change-Id: Ia0bb32fd92ae723c505d0fce5691e0fe0540f10d