Updated openstack/openstack Project: openstack/nova  08cc0d95a7746f6ba131fcec7a642caa4a3dc1ed
Require Placement 1.10 in nova-status upgrade check

The FilterScheduler is now using the GET /allocation_candidates
API from the Placement service which was added in microversion
1.10, and the nova-compute service is also going to rely on it
for updating allocations during move operations.

So this makes nova-status require Placement 1.10 as a minimum
for Pike. There are already release notes and some docs changes
mentioning that 1.10 is required for the scheduler and that
Placement should be upgraded before the scheduler, so another
release note is not added here.

Related to blueprint placement-claims

Change-Id: Iba0a734eb5535ec7cc394858ef15f6eebf8a9bde