Updated openstack/openstack Project: openstack/nova  f401d18a3d416a02156bc656717dfeed47dc217f
Improve stable-api doc with current API state

extensions, stevedore and extensions config options are gone
and plain router is introduced in Pike.

Also this documents was little bit confusing about current and
old state of APIs.

This commit makes it more clear and reflect the current API state
and also describe the evolution of APIs.

Change-Id: I6522100a78241400b1ea059a39a32c259fe6ab90