Updated openstack/openstack Project: openstack/neutron  a04f8fd049b50b1f09ec9ab55564bdceda36362d
DVR: Fix router_update failure when agent restarts

Router update task fails when agent restarts with DVR routers
as it was failing adding an IP rule to the namespace.

The IP rule matching code was not finding a match for
a rule for an interface since we are not specifying an
IP address, but the resulting rule does have the "any" IP
address in its output, for example, 0.0.0.0/0.  Change
to always supply the IP address.

Change-Id: Ic2e80ebb59ac9e0e0063e5f6e69f3d66abe775a1
Closes-Bug: #1702790