Updated openstack/openstack Project: openstack/neutron  b940f041a6427f9adc967d18b888c1f860552963
Reduce rally sub-port count

This test is executed 4 times and creating 1000 Neutron ports
just to run this scenario 4 times results in this job taking
15 minutes to complete its iterations.

This cuts the count in half to 125 per execution to cut the run
time in half and ensure we don't get to close to the gate timeout.

This was done once before as part of
817a19c4b9ced01af9bbb6f7383283e3f0f0e9bd but unfortunately being
part of another SLA change resulted in it being reverted as part
of the SLA change.

Change-Id: I61466d87b002252efc163cbb5d03eafc5d4da3fb