Updated openstack/openstack Project: openstack/nova  c58d4422d87af8b1e8583972d3eef5c7db327264
Remove provider allocs in confirm/revert resize

Now that the scheduler creates a doubled-up allocation for the duration
of a move operation (with part of the allocation referring to the
source and part referring to the destination host), we need to remove
the source provider when confirming the resize and remove the
destination provider from the allocation when reverting a resize. This
patch adds this logic in the RT's drop_move_claim() method.

Change-Id: I6f8afe6680f83125da9381c812016b3623503825
Co-Authored-By: Dan Smith <dms@danplanet.com>
Fixes-bug: #1707071