Updated openstack/openstack Project: openstack/nova  2918c70e1aff1c41d4bdaf0861cff0262b57f878
doc: add another up-call caveat for cells v2 for xenapi aggregates

There is an up-call from the xenapi driver when performing a live
migration and --block-migrate is not specified such that the driver
checks to see if the source and destination hosts are in the same
aggregate. This fails in a super-conductor setup because the
aggregates are now in the API database and the cell conductor
won't be able to access that database by design.

Change-Id: I6c880c72d87eb0116cb57371e5d600dced2915f7
Related-Bug: #1709594