Updated openstack/openstack Project: openstack/nova  4579d2e5573ae1bbabb51ee46ef26598d9410b15
doc: add superconductor up-call caveat for cross_az_attach=False

When attaching a volume, if CONF.cinder.cross_az_attach=False,
we attempt to lookup the AZ that the instance is in and compare
that to the volume AZ. The instance AZ lookup involves getting
host aggregates which are in the API database, and in superconductor
mode the cell conductor can't access to the API database.

For volume attach, we could remove this check from the compute
service since we already check this in nova-api so the check in
nova-compute is redundant.

However, we still have a problem with checking this during boot
from volume where nova-compute creates the volume and then
attaches it. At that point it's too late and we'll fail.

We should eventually create volumes during boot from volume in
conductor, like we've talked about doing with neutron ports
during server create, but that's not something we have today
so we need to point out this limitation.

Change-Id: I9cbe41e41f8ccdc9962ab593f313b5a47314a9d1