Updated openstack/openstack Project: openstack-infra/project-config  ce12a892ef2acc557d33e86138c221a651e87b07
Temporarily stop using rax-ord

In preparation for a mirror server replacement, temporarily cease
using rax-ord. This gives us an opportunity to allow htcacheclean to
complete on the proxy cache volume, allows us leeway to move the
Cinder volumes from the old mirror to its replacement, and also
stops the recurrence of pip hash mismatch errors in jobs while we
get the new server online. We're upgrading from 4GB to 8GB
performance flavor because we seem to be spending a fair amount of
time at or over the bandwidth limit for the current flavor.

Change-Id: I954fda2fecef0832e81511d07c8526036d2dabf9