Updated openstack/openstack Project: openstack/cinder  7f69969345ec027595d2530d646cc5fd820be65a
Make attachment_update set status to attaching

The new attachment_update method in Cinder's API creates an attachment
object and populates it with the provided connector info. In addition,
we set the volumes status to in-use and update the attachment object
status to "attached".

This isn't really accurate though, because we don't know if the volume
is actually attached (connected) by the consumer or not.  Also a big
side effect here is that currently all of our tests and automation
use volume-status to determine if a volume is fully connected/ready
for use and that everything went well.  It's used as an ack in most
cases.

This change goes back to using multiple states to signify where a
an attachment is in it's life-cycle:
1. reserved
     We've created an empty attachment record but haven't done anything
     with it yet.
2. attaching
     We provided a connector and set up the TGT so that everything is
     ready for a consumer to connect/use it.
3. in-use
     An ACK back from the consumer letting us know that they connected
     it successfully and are doing their thing.

Some consumers don't need or care about this last step, and we're going
to provide a means to set it straight to attached/in-use, but for
this bug we don't need to introduce that particular *feature*.

Sadly, this requires a micro-version bump and a new API call to
toggle the state from the API, pushing us to 3.44.

closes-bug #1710295

Change-Id: I57631d3deddb2d7cd244584e82206ee17fe2dd78